+++ src/DependencyInjection/UserExtension.php
@@ -14,12 +14,10 @@ declare(strict_types=1);
 namespace EightMarq\UserBundle\DependencyInjection;
 
 use EightMarq\CoreBundle\DependencyInjection\AbstractExtension;
-use EightMarq\UserBundle\Entity\User;
 use Exception;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
-use Symfony\Component\Security\Core\User\UserInterface;
 
 class UserExtension extends AbstractExtension
 {
@@ -38,15 +36,4 @@ class UserExtension extends AbstractExtension
 
         $loader->load('services.yaml');
     }
-
-    /**
-     * @param ContainerBuilder $container
-     */
-    public function prepend(ContainerBuilder $container): void
-    {
-        $this->targetEntities[UserInterface::class] = User::class;
-
-        parent::prepend($container);
-    }
-
 }