+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # Changelog
 
+## 0.6.0
+
+### Change
+
+* Modify `EightMarq\UserBundle\Entity\User` entity to a MappedSuperclass
+* Modify `EightMarq\UserBundle\Repository\UserRepository` to be more general
+* Remove `EightMarq\UserBundle\Entity\User` entity from Doctrine interface definition
+
 ## 0.5.0
 
 ### Added
 
-Added User entity
-Added User repository
-Added PasswordUpdater
+* Added `EightMarq\UserBundle\Entity\User` entity
+* Added `EightMarq\UserBundle\Repository\UserRepository` repository
+* Added `EightMarq\UserBundle\Security\PasswordUpdater`